refactor: remove drum application dependencies from musin library
- Move DrumpadConfig into musin/ui/drumpad_config.h; drum/config.h
  aliases it and keeps only the product values
- TempoHandler takes the tempo knob BPM range via
  set_tempo_control_range() instead of reading drum::config, and the
  RETRIGGER_SYNC_ON_PLAYBUTTON policy check moves to the caller in
  SequencerController::start()
- Move generic sysex::Chunk from drum/sysex/chunk.h to
  musin/midi/sysex_chunk.h so the MIDI input queue no longer includes
  drum headers
- Drop vestigial pizza_controls.h include from sequencer_controller.cpp
